+++ suppliers/cdopromo.py
